Build Your Own/requirements.txt
python-telegram-bot==13.15
pillow==11.3.0
opencv-python==4.12.0.88
psutil==7.0.0
pywin32==311
humanize==4.12.3
pyinstaller==6.14.2